requirements.txt
absl-py==2.4.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
beautifulsoup4==4.12.2
blinker==1.9.0
blis==1.3.3
boto3==1.42.66
botocore==1.42.66
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
coqpit-config==0.2.4
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.13
datasets==4.7.0
decorator==5.2.1
dill==0.4.0
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
filelock==3.25.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
ko-speech-tools==0.1.0
kokoro==0.7.16
language-tags==1.2.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
misaki==0.9.4
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==1.26.4
onnxruntime==1.24.3
packaging==26.0
pandas==3.0.1
pathvalidate==3.3.1
phonemizer-fork==3.3.2
pillow==12.1.1
piper-tts==1.4.1
platformdirs==4.9.4
pluggy==1.6.0
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.8.0
pysbd==0.3.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
segments==2.4.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
supertonic==1.1.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchcodec==0.10.0
tqdm==4.67.3
transformers==4.57.6
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
